portal/pyproject.toml
[project]
name = "portal"
version = "0.1.0"
description = "ToolBox Pro — Portal login + dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.36",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = []
